requirements.txt
pyqt6>=6.4
pyenchant>=3.0.0
fpdf2>=2.7.9
GitPython>=3.1.45